Finally figured out correct directory structure for the project, and placed tests in their own folder
