--- api/__init__.py
@@ -1,3 +1,0 @@
-# For relative imports to work in Python 3.6
-import os, sys
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
