+++ api/lib/singleton.py
@@ -1,0 +1,13 @@
+
+# This implementation is following the suggestion found here
+# https://stackoverflow.com/questions/6760685/creating-a-singleton-in-python
+
+# To use, do: "class MyClass (metaclass = Singleton)"
+
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
