--- api/test_files/normal_file.txt
@@ -1,5 +1,0 @@
-This is some text
-
-this is some more text
-
-bye!
