+++ launch_tests.py
@@ -1,0 +1,2 @@
+import tests.file_upload_test
+import tests.file_utils_test
