+++ tests/old/post_img
@@ -1,0 +1,12 @@
+#!/usr/bin/env bash
+
+if [ $# -ne 1 ] ; then
+	>&2 echo "$0: Must be run with file argument"
+	exit 1
+fi
+if [ ! -f "$1" ] ; then
+	>&2 echo "$1 is not a file"
+	exit 1
+fi
+
+curl -X POST -F "picture=@$1" localhost:3000/pictures
