+++ tests/test_files/normal_file.txt
@@ -1,0 +1,5 @@
+This is some text
+
+this is some more text
+
+bye!
